backend/requirements.txt
fastapi~=0.115.0
uvicorn[standard]~=0.30.0
SQLAlchemy[asyncio]~=2.0.25
alembic~=1.13.1
pydantic~=2.7.0
pyyaml~=6.0.1
asyncpg~=0.30.0

pytest~=8.0.0
pytest-asyncio~=0.23.0
pytest-cov~=5.0.0
ruff~=0.6.0
black~=24.4.2
httpx~=0.27.0
aiosqlite~=0.20.0

